Заменил int на long long

+++ Математика/Binpow.cpp
@@ -1,5 +1,5 @@
-int binpow(int a, int n) {
-    int res = 1;
+ll binpow(ll a, ll n) {
+    ll res = 1;
     while (n != 0) {
         if (n & 1)
             res = res * a;
